+++ eslint.config.js
@@ -67,6 +67,7 @@ export default [
 			'@typescript-eslint/no-unused-vars': [ 'error', { args: 'none' } ],
 			'@typescript-eslint/no-explicit-any': 'off',
 			'@typescript-eslint/no-empty-object-type': 'off',
+			'@typescript-eslint/no-unsafe-function-type': 'off',
 		},
 	},