+++ src/three/plugins/batched/BatchedTilesPlugin.js
@@ -1,6 +1,6 @@
 /** @import { WebGLRenderer, Material } from 'three' */
 import { WebGLArrayRenderTarget, MeshBasicMaterial, DataTexture, REVISION } from 'three';
-import { FullScreenQuad } from 'three/examples/jsm/postprocessing/Pass.js';
+import { FullScreenQuad } from 'three/addons/postprocessing/Pass.js';
 import { ExpandingBatchedMesh } from './ExpandingBatchedMesh.js';
 import { convertMapToArrayTexture, isColorWhite } from './utilities.js';