+++ src/three/renderer/tiles/TilesRenderer.js
@@ -20,7 +20,7 @@ import { TileBoundingVolume } from '../math/TileBoundingVolume.js';
 import { ExtendedFrustum } from '../math/ExtendedFrustum.js';
 import { estimateBytesUsed } from '../utils/MemoryUtils.js';
 import { WGS84_ELLIPSOID } from '../math/GeoConstants.js';
-import { GLTFLoader } from 'three/examples/jsm/loaders/GLTFLoader.js';
+import { GLTFLoader } from 'three/addons/loaders/GLTFLoader.js';
 
 const _mat = /* @__PURE__ */ new Matrix4();
 const _euler = /* @__PURE__ */ new Euler();
@@ -559,7 +559,7 @@ export class TilesRenderer extends TilesRendererBase {
 			tempMat.premultiply( camera.matrixWorldInverse );
 			tempMat.premultiply( camera.projectionMatrix );
 
-			frustum.setFromProjectionMatrix( tempMat );
+			frustum.setFromProjectionMatrix( tempMat, camera.coordinateSystem, camera.reversedDepth );
 
 			// get transform position in group root frame
 			position.set( 0, 0, 0 );